added new firebase code
