+++ package.json
@@ -10,6 +10,8 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "firebase": "^9.6.1",
+    "firebase-admin": "^11.5.0",
+    "firebaseui": "^6.0.2",
     "gh-pages": "^3.2.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
